Converted separate chat/console commands into 1 covalence one
Covalence commands have many advantages over regular commands: first and foremost, they can be registered under different aliases easily without using methods for redirects. Second, they can be registered with a set of permissions required, so there's no need to check for that permission later in the code of the method. Third, and this is the biggest bonus, they're completely agnostic when it comes to distinction between console and chat - same exact logic can be used.